Refactor code to become more readable, ODE-solver to PEtab file processesing
