+++ src/Create_PEtab_model.jl
@@ -3,7 +3,6 @@
                    forceBuildJuliaFiles::Bool=false,
                    verbose::Bool=true,
                    ifElseToEvent::Bool=true,
-                   jlFile::Bool=false,
                    jlFilePath::String="")::PEtabModel
 
 Parses a PEtab specified problem with a YAML-file located at `pathYAML` into a Julia-accessible format.
@@ -16,15 +15,14 @@ When parsing a PEtab problem, several things happen under the hood:
 
 All of this happens automatically, and resulting files are stored under `petabModel.dirJulia`. To save time, `forceBuildJlFiles=false` by default, which means that Julia files are not rebuilt if they already exist.
 
-In case a Julia model files is provided instead of a SBML file, set `jlFile` and provide the file path under `jlFilePath`.
+In case a Julia model files is provided instead of a SBML file provide file path under `jlFilePath`.
 
 # Arguments
 - `pathYAML::String`: Path to the PEtab problem YAML file.
 - `forceBuildJuliaFiles::Bool=false`: If `true`, forces the creation of Julia files for the problem even if they already exist.
 - `verbose::Bool=true`: If `true`, displays verbose output during parsing.
 - `ifElseToEvent::Bool=true`: If `true`, rewrites `if-else` statements in the SBML model as event-based callbacks.
-- `jlFile::Bool=false`: If `true`, loads a pre-existing Julia file for the problem instead of parsing from the YAML.
-- `jlFilePath::String=""`: Path to an existing Julia file to load if `jlFile` is `true`.
+- `jlFilePath::String=""`: Path to an existing Julia file. Should only be provided if a Julia model file is availble.
 
 # Example
 ```julia
@@ -35,137 +33,110 @@ function readPEtabModel(pathYAML::String;
                         forceBuildJuliaFiles::Bool=false,
                         verbose::Bool=true,
                         ifElseToEvent::Bool=true,
-                        jlFile::Bool=false,
                         jlFilePath::String="", 
                         customParameterValues::Union{Nothing, Dict}=nothing)::PEtabModel
 
+    jlFile = isempty(jlFilePath) ? false : true
+
     pathSBML, pathParameters, pathConditions, pathObservables, pathMeasurements, dirJulia, dirModel, modelName = readPEtabYamlFile(pathYAML, jlFile=jlFile)
 
-    verbose == true && printstyled("[ Info:", color=123, bold=true)
-    verbose == true && @printf(" Building PEtabModel for %s\n", modelName)
+    verbose == true && @info "Building PEtabModel for $modelName"
 
     if jlFile == false
-
         pathModelJlFile = joinpath(dirJulia, modelName * ".jl")
-
-        if !isfile(pathModelJlFile)
-            if verbose == true
-                printstyled("[ Info:", color=123, bold=true)
-                print(" Building Julia model file as it does not exist ...")
-            end
-            bBuild = @elapsed modelDict = XmlToModellingToolkit(pathSBML, pathModelJlFile, modelName, ifElseToEvent=ifElseToEvent)
-            verbose == true && @printf(" done. Time = %.1es\n", bBuild)
-
-        elseif isfile(pathModelJlFile) && forceBuildJuliaFiles == false && verbose == true
-            printstyled("[ Info:", color=123, bold=true)
-            print(" Julia model file exists and will not be rebuilt\n")
-
-        elseif forceBuildJuliaFiles == true
-            if verbose == true
-                printstyled("[ Info:", color=123, bold=true)
-                print(" By user option rebuilds Julia model file ...")
-            end
-            isfile(pathModelJlFile) == true && rm(pathModelJlFile)
+        if !isfile(pathModelJlFile) || forceBuildJuliaFiles == true
+            verbose == true && printstyled("[ Info:", color=123, bold=true)
+            verbose == true && forceBuildJuliaFiles && print(" By user option rebuilds Julia model file ...")
+            verbose == true && !forceBuildJuliaFiles && print(" Building Julia model file as it does not exist ...")
+                
             bBuild = @elapsed modelDict = XmlToModellingToolkit(pathSBML, pathModelJlFile, modelName, ifElseToEvent=ifElseToEvent)
             verbose == true && @printf(" done. Time = %.1es\n", bBuild)
         end
 
-    else
-        if jlFilePath == ""
-            throw("If jlFile=true you must specify jlFilePath")
-        elseif !isfile(jlFilePath)
-            throw("File" * jlFilePath * "does not exist")
-        else
-            modelDict, pathModelJlFile = JLToModellingToolkit(jlFilePath, dirJulia, modelName, ifElseToEvent=ifElseToEvent)
+        if isfile(pathModelJlFile) && forceBuildJuliaFiles == false
+            verbose == true && printstyled("[ Info:", color=123, bold=true)
+            verbose == true && print(" Julia model file exists and will not be rebuilt\n")
         end
+    end
+
+    if jlFile == true
+        if isempty(jlFilePath) || !isfile(jlFilePath)
+            @error "In case jlFile=true you must provide the path to a valid Julia model model. $jlFilePath is not a valid file"
+        end
+           
+        modelDict, pathModelJlFile = JLToModellingToolkit(jlFilePath, dirJulia, modelName, ifElseToEvent=ifElseToEvent)
     end
 
     addParameterForConditionSpecificInitialValues(pathModelJlFile, pathConditions, pathParameters)
 
     # Load model ODE-system
     @assert isfile(pathModelJlFile)
-    verbose == true && printstyled("[ Info:", color=123, bold=true)
+    verbose == true && printstyled("[ Info:", color=123, bold=true) 
     verbose == true && print(" Symbolically processes ODE-system ...")
-    bTake = @elapsed begin
+    timeTake = @elapsed begin
         _getODESystem = @RuntimeGeneratedFunction(Meta.parse(getFunctionsAsString(pathModelJlFile, 1)[1]))
         _odeSystem, stateMap, parameterMap = _getODESystem("https://xkcd.com/303/") # Argument needed by @RuntimeGeneratedFunction
         odeSystem = structural_simplify(_odeSystem)
-        # TODO : Make these to strings here to save conversions
         parameterNames = parameters(odeSystem)
         stateNames = states(odeSystem)
     end
-    verbose == true && @printf(" done. Time = %.1es\n", bTake)
+    verbose == true && @printf(" done. Time = %.1es\n", timeTake)
 
-    # Build functions for observables, sd and u0 if does not exist and include
     path_u0_h_sigma = joinpath(dirJulia, modelName * "_h_sd_u0.jl")
-    path_D_h_sd = joinpath(dirJulia, modelName * "_D_h_sd.jl")
     if !isfile(path_u0_h_sigma) || forceBuildJuliaFiles == true
-        if verbose == true && !isfile(path_u0_h_sigma)
-            printstyled("[ Info:", color=123, bold=true)
-            print(" Building u0, h σ file as it does not exist ...")
-        elseif verbose == true
-            printstyled("[ Info:", color=123, bold=true)
-            print(" By user option rebuilds u0, h σ file ...")
-        end
+        verbose == true && printstyled("[ Info:", color=123, bold=true)
+        verbose == true && !isfile(path_u0_h_sigma) && print(" Building u0, h and σ file as it does not exist ...")
+        verbose == true && isfile(path_u0_h_sigma) && print(" By user option rebuilds u0, h and σ file ...")
         if !@isdefined(modelDict)
             modelDict = XmlToModellingToolkit(pathSBML, pathModelJlFile, modelName, writeToFile=false, ifElseToEvent=ifElseToEvent)
         end
         bBuild = @elapsed create_σ_h_u0_File(modelName, pathYAML, dirJulia, odeSystem, parameterMap, stateMap, modelDict, jlFile=jlFile, customParameterValues=customParameterValues)
         verbose == true && @printf(" done. Time = %.1es\n", bBuild)
-    elseif verbose == true
-        printstyled("[ Info:", color=123, bold=true)
-        print(" u0, h and σ file exists and will not be rebuilt\n")
+    else
+        verbose == true && printstyled("[ Info:", color=123, bold=true) 
+        verbose == true && print(" u0, h and σ file exists and will not be rebuilt\n")
     end
-
-    if !isfile(path_D_h_sd) || forceBuildJuliaFiles == true
-        if verbose == true && !isfile(path_D_h_sd)
-            printstyled("[ Info:", color=123, bold=true)
-            print(" Building ∂h∂p, ∂h∂u, ∂σ∂p and ∂σ∂u file as it does not exist ...")
-        elseif verbose == true
-            printstyled("[ Info:", color=123, bold=true)
-            print(" By user option rebuilds ∂h∂p, ∂h∂u, ∂σ∂p and ∂σ∂u file ...")
-        end
-        if !@isdefined(modelDict)
-            modelDict = XmlToModellingToolkit(pathSBML, pathModelJlFile, modelName, writeToFile=false, ifElseToEvent=ifElseToEvent)
-        end
-        bBuild = @elapsed createDerivative_σ_h_File(modelName, pathYAML, dirJulia, odeSystem, parameterMap, stateMap, modelDict, jlFile=jlFile, customParameterValues=customParameterValues)
-        verbose == true && @printf(" done. Time = %.1es\n", bBuild)
-    elseif verbose == true
-        printstyled("[ Info:", color=123, bold=true)
-        print(" ∂h∂p, ∂h∂u, ∂σ∂p and ∂σ∂u file exists and will not be rebuilt\n")
-    end
-
     @assert isfile(path_u0_h_sigma)
     h_u0_σ_Functions = getFunctionsAsString(path_u0_h_sigma, 4)
     compute_h = @RuntimeGeneratedFunction(Meta.parse(h_u0_σ_Functions[1]))
     compute_u0! = @RuntimeGeneratedFunction(Meta.parse(h_u0_σ_Functions[2]))
     compute_u0 = @RuntimeGeneratedFunction(Meta.parse(h_u0_σ_Functions[3]))
     compute_σ = @RuntimeGeneratedFunction(Meta.parse(h_u0_σ_Functions[4]))
+
+    path_D_h_sd = joinpath(dirJulia, modelName * "_D_h_sd.jl")
+    if !isfile(path_D_h_sd) || forceBuildJuliaFiles == true
+        verbose == true && printstyled("[ Info:", color=123, bold=true)
+        verbose == true && !isfile(path_u0_h_sigma) && print(" Building ∂h∂p, ∂h∂u, ∂σ∂p and ∂σ∂u file as it does not exist ...")
+        verbose == true && isfile(path_u0_h_sigma) && print(" By user option rebuilds ∂h∂p, ∂h∂u, ∂σ∂p and ∂σ∂u file ...")
+        if !@isdefined(modelDict)
+            modelDict = XmlToModellingToolkit(pathSBML, pathModelJlFile, modelName, writeToFile=false, ifElseToEvent=ifElseToEvent)
+        end
+        bBuild = @elapsed createDerivative_σ_h_File(modelName, pathYAML, dirJulia, odeSystem, parameterMap, stateMap, modelDict, jlFile=jlFile, customParameterValues=customParameterValues)
+        verbose == true && @printf(" done. Time = %.1es\n", bBuild)
+    else verbose == true
+        verbose == true && printstyled("[ Info:", color=123, bold=true) 
+        verbose == true && print(" ∂h∂p, ∂h∂u, ∂σ∂p and ∂σ∂u file exists and will not be rebuilt\n")
+    end
     @assert isfile(path_D_h_sd)
     ∂_h_σ_Functions = getFunctionsAsString(path_D_h_sd, 4)
     compute_∂h∂u! = @RuntimeGeneratedFunction(Meta.parse(∂_h_σ_Functions[1]))
     compute_∂h∂p! = @RuntimeGeneratedFunction(Meta.parse(∂_h_σ_Functions[2]))
     compute_∂σ∂σu! = @RuntimeGeneratedFunction(Meta.parse(∂_h_σ_Functions[3]))
     compute_∂σ∂σp! = @RuntimeGeneratedFunction(Meta.parse(∂_h_σ_Functions[4]))
-
+        
     pathCallback = joinpath(dirJulia, modelName * "_callbacks.jl")
     if !isfile(pathCallback) || forceBuildJuliaFiles == true
-        if verbose == true && !isfile(pathCallback)
-            printstyled("[ Info:", color=123, bold=true)
-            print(" Building callback file as it does not exist ...")
-        elseif verbose == true
-            printstyled("[ Info:", color=123, bold=true)
-            print(" By user option rebuilds callback file ...")
-        end
+        verbose == true && printstyled("[ Info:", color=123, bold=true)
+        verbose == true && !isfile(pathCallback) && print(" Building callback file as it does not exist ...")
+        verbose == true && isfile(pathCallback) && print(" By user option rebuilds callback file ...")
         if !@isdefined(modelDict)
             modelDict = XmlToModellingToolkit(pathSBML, pathModelJlFile, modelName, writeToFile=false, ifElseToEvent=ifElseToEvent)
         end
         bBuild = @elapsed createCallbacksForTimeDepedentPiecewise(odeSystem, parameterMap, stateMap, modelDict, modelName, pathYAML, dirJulia, jlFile = jlFile, customParameterValues=customParameterValues)
         verbose == true && @printf(" done. Time = %.1es\n", bBuild)
-
-    elseif verbose == true
-        printstyled("[ Info:", color=123, bold=true)
-        print(" Callback file exists and will not be rebuilt\n")
+    else
+        verbose == true && printstyled("[ Info:", color=123, bold=true) 
+        verbose == true && print(" Callback file exists and will not be rebuilt\n")
     end
     @assert isfile(pathCallback)
     strGetCallbacks = getFunctionsAsString(pathCallback, 2)
@@ -387,22 +358,4 @@ function startsWithx(str, x)
         return true
     end
     return false
-end
-
-
-import Base.show
-function show(io::IO, a::PEtabModel)
-
-    modelName = @sprintf("%s", a.modelName)
-    numberOfODEStates = @sprintf("%d", length(a.stateNames))
-    numberOfODEParameters = @sprintf("%d", length(a.parameterNames))
-
-    printstyled("PEtabModel", color=116)
-    print(" for model ")
-    printstyled(modelName, color=116)
-    print(". ODE-system has ")
-    printstyled(numberOfODEStates * " states", color=116)
-    print(" and ")
-    printstyled(numberOfODEParameters * " parameters.", color=116)
-    @printf("\nGenerated Julia files are at %s", a.dirJulia)
 end
