+++ src/Process_PEtab_files/Process_parameters.jl
@@ -47,24 +47,26 @@ function processParameters(parametersFile::CSV.File; customParameterValues::Unio
 
         # In some case when working with the model the user might want to change model parameters but not go the entire 
         # way to the PEtab-files. This ensure ParametersInfo gets its parameters correct.
-        if !isnothing(customParameterValues)
-            keysDict = collect(keys(customParameterValues))
-            iKey = findfirst(x -> x == parameterId[i], keysDict)
-            isnothing(iKey) && continue
-            valueChangeTo = customParameterValues[keysDict[iKey]]
-            if typeof(valueChangeTo) <: Real
-                estimate[i] = false
-                nominalValue[i] = valueChangeTo
-            elseif isNumber(valueChangeTo)
-                estimate[i] = false
-                nominalValue[i] = parse(Float64, valueChangeTo)
-            elseif valueChangeTo == "estimate"
-                estimate[i] = true
-            else
-                PEtabFileError("For PEtab select a parameter must be set to either a estimate or a number not $valueChangeTo")
-            end
+        if isnothing(customParameterValues) 
+            continue
+        end
+        keysDict = collect(keys(customParameterValues))
+        iKey = findfirst(x -> x == parameterId[i], keysDict)
+        isnothing(iKey) && continue
+        valueChangeTo = customParameterValues[keysDict[iKey]]
+        if typeof(valueChangeTo) <: Real
+            estimate[i] = false
+            nominalValue[i] = valueChangeTo
+        elseif isNumber(valueChangeTo)
+            estimate[i] = false
+            nominalValue[i] = parse(Float64, valueChangeTo)
+        elseif valueChangeTo == "estimate"
+            estimate[i] = true
+        else
+            PEtabFileError("For PEtab select a parameter must be set to either a estimate or a number not $valueChangeTo")
         end
     end
+
     nParametersToEstimate::Int64 = Int64(sum(estimate))
     return ParametersInfo(nominalValue, lowerBound, upperBound, parameterId, parameterScale, estimate, nParametersToEstimate)
 end
