+++ src/Solve_ODE/Solve_for_steady_state.jl
@@ -1,50 +1,11 @@
-"""
-    SteadyStateSolverOptions(method::Symbol;
-                             howCheckSimulationReachedSteadyState::Symbol=:wrms,
-                             rootfindingAlgorithm=nothing,
-                             abstol=nothing, 
-                             reltol=nothing, 
-                             maxiters=nothing)
-
-Setup options for finding steady-state via either `method=:Rootfinding` or `method=:Simulate`.
-
-For `method=:Rootfinding`, the steady-state `u*` is found by solving the problem `du = f(u, p, t) ≈ 0` with tolerances `abstol` and `reltol` via an automatically chosen optimization algorithm (`rootfindingAlgorithm=nothing`) or via any algorithm in NonlinearSolve.jl. 
-
-For `method=:Simulate`, the steady-state `u*` is found by simulating the ODE system until `du = f(u, p, t) ≈ 0`. Two options are available for `howCheckSimulationReachedSteadyState`:
-- `:wrms` : Weighted root-mean square √(∑((du ./ (reltol * u .+ abstol)).^2) / length(u)) < 1
-- `:Newton` : If Newton-step `Δu` is sufficiently small √(∑((Δu ./ (reltol * u .+ abstol)).^2) / length(u)) < 1. 
-        - Newton often performs better but requires an invertible Jacobian. In case it's not fulfilled, the code switches automatically to `:wrms`.    
-
-`maxiters` refers to either the maximum number of rootfinding steps or the maximum number of integration steps, depending on the chosen method.        
-"""
-function SteadyStateSolverOptions(method::Symbol;
-                                  howCheckSimulationReachedSteadyState::Symbol=:wrms,
-                                  rootfindingAlgorithm::Union{Nothing, NonlinearSolve.AbstractNonlinearSolveAlgorithm}=nothing,
-                                  abstol=nothing, 
-                                  reltol=nothing, 
-                                  maxiters::Union{Nothing, Int64}=nothing)::SteadyStateSolverOptions
-
-    @assert method ∈ [:Rootfinding, :Simulate] "Method used to find steady state can either be :Rootfinding or :Simulate not $method"
-    
-    if method === :Simulate
-        return _getSteadyStateSolverOptions(howCheckSimulationReachedSteadyState, abstol, reltol, maxiters)
-    else
-        return _getSteadyStateSolverOptions(rootfindingAlgorithm, abstol, reltol, maxiters)
-    end
-end
-
-
 function _getSteadyStateSolverOptions(howCheckSimulationReachedSteadyState::Symbol, 
                                       abstol,
                                       reltol, 
                                       maxiters)::SteadyStateSolverOptions
 
-    _abstol = isnothing(abstol) ? 1e-8 : abstol
-    _reltol = isnothing(reltol) ? 1e-8 : reltol
-    _maxiters = isnothing(maxiters) ? Int64(1e4) : maxiters                                      
-
     @assert howCheckSimulationReachedSteadyState ∈ [:Newton, :wrms] "When steady states are computed via simulations howCheckSimulationReachedSteadyState must be :wrms or :Newton not $howCheckSimulationReachedSteadyState"
-    return SteadyStateSolverOptions(:Simulate, nothing, howCheckSimulationReachedSteadyState, _abstol, _reltol, _maxiters, nothing, nothing)
+
+    return SteadyStateSolverOptions(:Simulate, nothing, howCheckSimulationReachedSteadyState, abstol, reltol, maxiters, nothing, nothing)
 end
 function _getSteadyStateSolverOptions(rootfindingAlgorithm::Union{Nothing, NonlinearSolve.AbstractNonlinearSolveAlgorithm},
                                       abstol,
@@ -53,14 +14,9 @@ function _getSteadyStateSolverOptions(rootfindingAlgorithm::Union{Nothing, Nonli
 
     # Sanity check user input 
     @assert typeof(rootfindingAlgorithm) <: Union{Nothing, NonlinearSolve.AbstractNonlinearSolveAlgorithm} "When steady states are computed via rootfinding rootfindingAlgorithm must be nothing or a NonlinearSolveAlgorithm (https://docs.sciml.ai/NonlinearSolve/stable/solvers/NonlinearSystemSolvers/)"
-
-    _abstol = isnothing(abstol) ? 1e-8 : abstol
-    _reltol = isnothing(reltol) ? 1e-8 : reltol
-    _maxiters = isnothing(maxiters) ? Int64(1e4) : maxiters
-
+    
     _rootfindingAlgorithm = isnothing(rootfindingAlgorithm) ? NonlinearSolve.TrustRegion() : rootfindingAlgorithm
-
-    return SteadyStateSolverOptions(:Rootfinding, _rootfindingAlgorithm, :nothing, _abstol, _reltol, _maxiters, nothing, nothing)                                      
+    return SteadyStateSolverOptions(:Rootfinding, _rootfindingAlgorithm, :nothing, abstol, reltol, maxiters, nothing, nothing)                                      
 end
 function _getSteadyStateSolverOptions(ssSolverOptions::SteadyStateSolverOptions,
                                       odeProblem::ODEProblem, 
@@ -101,7 +57,6 @@ function _getSteadyStateSolverOptions(ssSolverOptions::SteadyStateSolverOptions,
 end
 
 
-# Solve ODE steady state 
 function solveODEPreEqulibrium!(uAtSS::AbstractVector,
                                 uAtT0::AbstractVector,
                                 odeProblem::ODEProblem,
@@ -119,10 +74,11 @@ function solveODEPreEqulibrium!(uAtSS::AbstractVector,
             uAtT0 .= odeSolution.prob.u0
         end
         return odeSolution
+    end
     
-    elseif ssSolverOptions.method === :Rootfinding
+    if ssSolverOptions.method === :Rootfinding
         
-        rootSolution = rootfindToSS(odeProblem, changeExperimentalCondition!, preEquilibrationId, ssSolverOptions)
+        rootSolution = rootfindSS(odeProblem, changeExperimentalCondition!, preEquilibrationId, ssSolverOptions)
         if rootSolution.retcode == ReturnCode.Success
             uAtSS .= rootSolution.u
             uAtT0 .= rootSolution.prob.u0
@@ -141,30 +97,13 @@ function simulateToSS(odeProblem::ODEProblem,
                       convertTspan::Bool)::ODESolution where S<:SciMLAlgorithm
 
     changeExperimentalCondition!(odeProblem.p, odeProblem.u0, preEquilibrationId)
-    _odeProblem = remakeODEProblemPreSolve(odeProblem, Inf, odeSolverOptions.solver, convertTspan)          
+    _odeProblem = setTspanODEProblem(odeProblem, Inf, odeSolverOptions.solver, convertTspan)          
 
     abstol, reltol, maxiters = odeSolverOptions.abstol, odeSolverOptions.reltol, odeSolverOptions.maxiters
     callbackSS = ssSolverOptions.callbackSS
 
     sol = solve(_odeProblem, solver, abstol=abstol, reltol=reltol, maxiters=maxiters, dense=false, callback=callbackSS)
     return sol
-end
-
-
-function rootfindToSS(odeProblem::ODEProblem,
-                      changeExperimentalCondition!::Function,
-                      preEquilibrationId::Symbol,
-                      ssSolverOptions::SteadyStateSolverOptions)::SciMLBase.NonlinearSolution
-
-    nonlinearProblem = remake(ssSolverOptions.nonlinearSolveProblem, u0=odeProblem.u0[:], p=odeProblem.p[:])
-    changeExperimentalCondition!(nonlinearProblem.p, nonlinearProblem.u0, preEquilibrationId)
-    nonlinearSolution = solve(nonlinearProblem, 
-                              ssSolverOptions.rootfindingAlgorithm, 
-                              abstol=ssSolverOptions.abstol, 
-                              reltol=ssSolverOptions.reltol, 
-                              maxiters=ssSolverOptions.maxiters)           
-
-    return nonlinearSolution
 end
 
 
@@ -207,38 +146,18 @@ function affectTerminateSS!(integrator)
 end
 
 
-import Base.show
-function show(io::IO, a::SteadyStateSolverOptions)
+function rootfindSS(odeProblem::ODEProblem,
+                    changeExperimentalCondition!::Function,
+                    preEquilibrationId::Symbol,
+                    ssSolverOptions::SteadyStateSolverOptions)::SciMLBase.NonlinearSolution
 
-    printstyled("SteadyStateSolverOptions", color=116)
-    if a.method === :Simulate
-        print(" with method ")
-        printstyled(":Simulate", color=116)
-        print(" ODE-model until du = f(u, p, t) ≈ 0.")
-        if a.howCheckSimulationReachedSteadyState === :wrms
-            @printf("\nSimulation terminated if wrms fulfill;\n√(∑((du ./ (reltol * u .+ abstol)).^2) / length(u)) < 1\n")
-        else
-            @printf("\nSimulation terminated if Newton-step Δu fulfill;\n√(∑((Δu ./ (reltol * u .+ abstol)).^2) / length(u)) < 1\n")
-        end
-        if isnothing(a.abstol)        
-            @printf("with (abstol, reltol) = default values.")
-        else
-            @printf("with (abstol, reltol) = (%.1e, %.1e)", a.abstol, a.reltol)
-        end
-    end
+    nonlinearProblem = remake(ssSolverOptions.nonlinearSolveProblem, u0=odeProblem.u0[:], p=odeProblem.p[:])
+    changeExperimentalCondition!(nonlinearProblem.p, nonlinearProblem.u0, preEquilibrationId)
+    nonlinearSolution = solve(nonlinearProblem, 
+                              ssSolverOptions.rootfindingAlgorithm, 
+                              abstol=ssSolverOptions.abstol, 
+                              reltol=ssSolverOptions.reltol, 
+                              maxiters=ssSolverOptions.maxiters)           
 
-    if a.method === :Rootfinding
-        print(" with method ")
-        printstyled(":Rootfinding", color=116)
-        print(" to solve du = f(u, p, t) ≈ 0.")
-        if isnothing(a.rootfindingAlgorithm)
-            @printf("\nAlgorithm : NonlinearSolve's heruistic. Options ")
-        else
-            algStr = string(a.rootfindingAlgorithm)
-            iEnd = findfirst(x -> x == '{', algStr)
-            algStr = algStr[1:iEnd-1] * "()"
-            @printf("\nAlgorithm : %s. Options ", algStr)
-        end
-        @printf("(abstol, reltol, maxiters) = (%.1e, %.1e, %d)", a.abstol, a.reltol, a.maxiters)
-    end
+    return nonlinearSolution
 end
